+++ src/la_lapack.F90
@@ -64,14 +64,14 @@ module la_lapack
           !>   \hline
           !>   B_{21} & B_{22} & 0 & 0 \\
           !>   0      & 0      & 0 & I \\
-          !>   C      & -S    & 0 & 0 
+          !>   C      & -S    & 0 & 0
           !> \end{array}
           !> \right] =
           !> \left[
           !> \begin{array}{c|c}
           !>   U_1 & 0 \\
           !>   \hline
-          !>   0 & U_2 
+          !>   0 & U_2
           !> \end{array}
           !> \right]
           !> \left[
@@ -79,14 +79,14 @@ module la_lapack
           !>   0 & 0 & -I & 0 \\
           !>   \hline
           !>   S & C & 0 & 0 \\
-          !>   0 & 0 & 0 & I 
+          !>   0 & 0 & 0 & I
           !> \end{array}
           !> \right]
           !> \left[
           !> \begin{array}{c|c}
           !>   V_1^H & 0 \\
           !>   \hline
-          !>   0 & V_2^H 
+          !>   0 & V_2^H
           !> \end{array}
           !> \right]
           !> \f]
@@ -4495,7 +4495,7 @@ module la_lapack
           !> (stopping criterion). Few tuning parameters (marked by [TP]) are
           !> available for the implementer.
           !> Further Details
-          !> 
+          !>
           !> GSVJ1 applies few sweeps of Jacobi rotations in the column space of
           !> the input M-by-N matrix A. The pivot pairs are taken from the (1,2)
           !> off-diagonal block in the corresponding N-by-N Gram matrix A^T * A. The